Fix wrong buffer size
timeval structure was copied using the size of a pointer,
not the size of a structure. Thankfully, the used size was
shorter than the destination buffer, and the pointer size
on x86_64 systems equals the size of `struct timeval`.

Nonetheless, the bug is fixed now, and the proper size
is used.

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>