libfru: Fix max output data length
The data encoded in BCD+, when decoded, may occupy twice as much
as the specified length of the encoded field. Prevously, the
`FRU_FIELDMAXARRAY` macro was defined to just a half of that.
The macro is used to define the size of the `decoded_field_t`
type introduced earlier, so the BCD+ encoded data over 31
characters would be rejected and decoding would fail on them.

That is fixed now.

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>