libfru, frugen: Fix field type preservation
This is a major refactoring to allow for the data field encoding type
to be preserved in the output, be it json or binary format.

Previously, the same `fru_field_t` type was used both for encoded
and decoded (exploded) fields. Recently, a new type, `decoded_field_t`,
was introduced for decoded fields. This commit makes use of that and
refactors all the encoding and decoding machinery.

Previously, the mandatory fields would be stored as source strings
and some hackery would be applied to specify encoding type for them.
On the other hands, custom fields would be right away encoded during
template reading (or command line processing) and would have encoding
type specified as part of the encoding process.

Now, all the input data (be it from template or from command line)
is stored in the new `decoded_field_t` structures and then uniformly
encoded using the refactored encoders.

That finally allows for using frugen to take a binary template,
modify some values via command line options, and write the data
into another (or the same) binary file without losing the encoding
type. The unsupported multirecord area record types are also
preserved that way, although not yet decodable into json or
encodable from it.

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>