frugen: Fix board date/time conversion
Decoder was using `gmtime` at an already UTC time value,
so it was applying a wrong offset and the time was displayed
in UTC-TZ instead of UTC+TZ.

Add a common conversion function to use both in json and
plain text output formatters, fix the TZ conversion algorithm.

Now the decoder always shows local time (as the encoder expects).

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>