ci: Adjust buildenv and cmake for Fedora
Fedora doesn't have a static version of libjson-c,
so add building of it into Dockerfile.fedora.

Also adjust CMakeLists.txt for easier future modification
that would allow for detecting whether or not the
static version of json-c is present. For now comment out
that code because it didn't work straight away for all
target distributions

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>